+++ gateway/gateway_test.go
@@ -12,8 +12,11 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"reflect"
+	"strconv"
 	"testing"
 	"time"
+
+	kithelper "github.com/rudderlabs/rudder-go-kit/testhelper"
 
 	"github.com/golang/mock/gomock"
 	"github.com/google/uuid"
@@ -350,10 +353,18 @@ var _ = Describe("Gateway", func() {
 				w.WriteHeader(http.StatusOK)
 				_, _ = io.WriteString(w, "OK")
 			}))
-			serverURL = whServer.URL
-			parsedURL, err := url.Parse(serverURL)
+			WHURL := whServer.URL
+			parsedURL, err := url.Parse(WHURL)
 			Expect(err).To(BeNil())
-			config.Set("Warehouse.webPort", parsedURL.Port())
+			whPort := parsedURL.Port()
+			GinkgoT().Setenv("RSERVER_WAREHOUSE_WEB_PORT", whPort)
+
+			serverPort, err := kithelper.GetFreePort()
+			Expect(err).To(BeNil())
+			serverURL = fmt.Sprintf("http://localhost:%d", serverPort)
+			GinkgoT().Setenv("RSERVER_GATEWAY_WEB_PORT", strconv.Itoa(serverPort))
+
+			loadConfig()
 
 			gateway = &HandleT{}
 			err = gateway.Setup(context.Background(), c.mockApp, c.mockBackendConfig, c.mockJobsDB, nil, c.mockVersionHandler, rsources.NewNoOpService(), sourcedebugger.NewNoOpService())
@@ -381,7 +392,6 @@ var _ = Describe("Gateway", func() {
 		}
 		verifyEndpoint := func(endpoints []string, method string) {
 			client := &http.Client{}
-			// baseURL := "http://localhost:8080"
 			for _, ep := range endpoints {
 				url := fmt.Sprintf("%s%s", serverURL, ep)
 				var req *http.Request
@@ -420,11 +430,14 @@ var _ = Describe("Gateway", func() {
 				close(wait)
 			}()
 			Eventually(func() bool {
-				resp, _ := http.Get(fmt.Sprintf("%s/version", serverURL))
+				resp, err := http.Get(fmt.Sprintf("%s/version", serverURL))
+				if err != nil {
+					return false
+				}
 				return resp.StatusCode == http.StatusOK
 			}, time.Second*10, time.Second).Should(BeTrue())
 
-			getEndpoint, postEndpoints, deleteEndpoints := getEndpointMethodMap()
+			getEndpoint, postEndpoints, deleteEndpoints := endpointsToVerify()
 			verifyEndpoint(getEndpoint, http.MethodGet)
 			verifyEndpoint(postEndpoints, http.MethodPost)
 			verifyEndpoint(deleteEndpoints, http.MethodDelete)
@@ -987,43 +1000,6 @@ var _ = Describe("Gateway", func() {
 		})
 	})
 
-	Context("Warehouse proxy", func() {
-		DescribeTable("forwarding requests to warehouse with different response codes",
-			func(url string, code int, payload string) {
-				gateway := &HandleT{}
-				whMock := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-					Expect(r.URL.String()).To(Equal(url))
-					Expect(r.Body)
-					Expect(r.Body).To(Not(BeNil()))
-					defer func() { _ = r.Body.Close() }()
-					reqBody, err := io.ReadAll(r.Body)
-					Expect(err).To(BeNil())
-					Expect(string(reqBody)).To(Equal(payload))
-					w.WriteHeader(code)
-				}))
-				GinkgoT().Setenv("WAREHOUSE_URL", whMock.URL)
-				GinkgoT().Setenv("RSERVER_WAREHOUSE_MODE", config.OffMode)
-				err := gateway.Setup(context.Background(), c.mockApp, c.mockBackendConfig, c.mockJobsDB, nil, c.mockVersionHandler, rsources.NewNoOpService(), sourcedebugger.NewNoOpService())
-				Expect(err).To(BeNil())
-
-				defer func() {
-					err := gateway.Shutdown()
-					Expect(err).To(BeNil())
-					whMock.Close()
-				}()
-
-				req := httptest.NewRequest("POST", "http://rudder-server"+url, bytes.NewBufferString(payload))
-				w := httptest.NewRecorder()
-				gateway.whProxy.ServeHTTP(w, req)
-				resp := w.Result()
-				Expect(resp.StatusCode).To(Equal(code))
-			},
-			Entry("successful request", "/v1/warehouse/pending-events", http.StatusOK, `{"source_id": "1", "task_run_id":"2"}`),
-			Entry("failed request", "/v1/warehouse/pending-events", http.StatusBadRequest, `{"source_id": "3", "task_run_id":"4"}`),
-			Entry("request with query parameters", "/v1/warehouse/pending-events?triggerUpload=true", http.StatusOK, `{"source_id": "5", "task_run_id":"6"}`),
-		)
-	})
-
 	Context("jobDataFromRequest", func() {
 		var (
 			gateway      *HandleT
@@ -1161,19 +1137,19 @@ func expectHandlerResponse(handler http.HandlerFunc, req *http.Request, response
 }
 
 // return all endpoints as key and method as value
-func getEndpointMethodMap() (getEndpoints, postEndpoints, deleteEndpoints []string) {
-	getEndpoints = []string{
+func endpointsToVerify() ([]string, []string, []string) {
+	getEndpoints := []string{
 		"/version",
 		"/robots.txt",
 		"/pixel/v1/track",
 		"/pixel/v1/page",
-		"/v1/warehouse",
 		"/v1/webhook",
 		"/v1/job-status/123",
 		"/v1/job-status/123/failed-records",
+		"/v1/warehouse/jobs/status",
 	}
 
-	postEndpoints = []string{
+	postEndpoints := []string{
 		"/v1/batch",
 		"/v1/identify",
 		"/v1/track",
@@ -1185,16 +1161,18 @@ func getEndpointMethodMap() (getEndpoints, postEndpoints, deleteEndpoints []stri
 		"/v1/import",
 		"/v1/audiencelist",
 		"/v1/webhook",
-		"/v1/warehouse",
 		"/beacon/v1/batch",
 		"/internal/v1/extract",
 		"/v1/warehouse/pending-events",
+		"/v1/warehouse/trigger-upload",
+		"/v1/warehouse/jobs",
+		"/v1/warehouse/fetch-tables",
 	}
 
-	deleteEndpoints = []string{
+	deleteEndpoints := []string{
 		"/v1/job-status/1234",
 	}
-	return
+	return getEndpoints, postEndpoints, deleteEndpoints
 }
 
 func allHandlers(gateway *HandleT) map[string]http.HandlerFunc {