+++ warehouse/upload.go
@@ -105,6 +105,7 @@ type UploadJob struct {
 	refreshPartitionBatchSize int
 	retryTimeWindow           time.Duration
 	minRetryAttempts          int
+	DisableAlter              bool
 
 	errorHandler ErrorHandler
 }
@@ -196,6 +197,7 @@ func (f *UploadJobFactory) NewUploadJob(ctx context.Context, dto *model.UploadJo
 		refreshPartitionBatchSize: config.GetInt("Warehouse.refreshPartitionBatchSize", 100),
 		retryTimeWindow:           retryTimeWindow,
 		minRetryAttempts:          minRetryAttempts,
+		DisableAlter:              config.GetBool("Warehouse.disableAlter", false),
 
 		alertSender: alerta.NewClient(
 			config.GetString("ALERTA_URL", "https://alerta.rudderstack.com/api/"),
@@ -794,6 +796,19 @@ func (job *UploadJob) UpdateTableSchema(tName string, tableSchemaDiff warehouseu
 }
 
 func (job *UploadJob) alterColumnsToWarehouse(tName string, columnsMap model.TableSchema) error {
+	if job.DisableAlter {
+		pkgLogger.Debugw("skipping alter columns to warehouse",
+			logfield.SourceID, job.warehouse.Source.ID,
+			logfield.SourceType, job.warehouse.Source.SourceDefinition.Name,
+			logfield.DestinationID, job.warehouse.Destination.ID,
+			logfield.DestinationType, job.warehouse.Destination.DestinationDefinition.Name,
+			logfield.WorkspaceID, job.warehouse.WorkspaceID,
+			logfield.TableName, tName,
+			"columns", columnsMap,
+		)
+		return nil
+	}
+
 	var responseToAlerta []model.AlterTableResponse
 	var errs []error